+++ TravelTipsAPI/ViewModels/db_basic/PreferRoutePostViewModel.cs
@@ -5,8 +5,8 @@ namespace TravelTipsAPI.ViewModels.db_basic
     public class PreferRoutePostViewModel
     {
         public int Type { get; set; }
-        public int DepartOsmId { get; set; }
-        public int ArrivalOsmId { get; set; }
+        public long DepartOsmId { get; set; }
+        public long ArrivalOsmId { get; set; }
         public int EstimateTime { get; set; }
         public int? LinkId { get; set; }
